Refactored code. Extracted register classes to separate package.
